renderer: C++ compatibility for tr_init

+++ code/renderer/tr_init.c
@@ -1850,7 +1850,7 @@ void R_Init( void ) {
 	// init function tables
 	//
 	for ( i = 0; i < FUNCTABLE_SIZE; i++ ) {
-		tr.sinTable[i] = sin( DEG2RAD( i * 360.0f / FUNCTABLE_SIZE ) + 0.0001f );
+		tr.sinTable[i] = sin( (double)(DEG2RAD( i * 360.0f / FUNCTABLE_SIZE ) + 0.0001f) );
 		tr.squareTable[i] = (i < FUNCTABLE_SIZE / 2) ? 1.0f : -1.0f;
 		if ( i == 0 ) {
 			tr.sawToothTable[i] = EPSILON;
